Add InstanceUploader class and pull out methods for GSheets
