+++ collect_app/src/main/java/org/odk/collect/android/upload/InstanceServerUploader.java
@@ -1,8 +1,6 @@
 package org.odk.collect.android.upload;
 
-import android.content.ContentValues;
 import android.content.SharedPreferences;
-import android.database.Cursor;
 import android.net.Uri;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
@@ -11,12 +9,10 @@ import com.google.android.gms.analytics.HitBuilders;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.dto.Instance;
 import org.odk.collect.android.http.HttpHeadResult;
 import org.odk.collect.android.http.OpenRosaHttpInterface;
 import org.odk.collect.android.preferences.PreferenceKeys;
-import org.odk.collect.android.provider.InstanceProviderAPI;
 import org.odk.collect.android.upload.result.SubmissionUploadAuthRequested;
 import org.odk.collect.android.upload.result.SubmissionUploadFatalError;
 import org.odk.collect.android.upload.result.SubmissionUploadNonFatalError;
@@ -40,9 +36,7 @@ import javax.net.ssl.HttpsURLConnection;
 
 import timber.log.Timber;
 
-import static org.odk.collect.android.provider.FormsProviderAPI.FormsColumns.AUTO_DELETE;
-
-public class InstanceServerUploader {
+public class InstanceServerUploader extends InstanceUploader {
     private static final String URL_PATH_SEP = "/";
     private static final String FAIL = "Error: ";
 
@@ -321,43 +315,5 @@ public class InstanceServerUploader {
         }
 
         return serverBase + submissionPath;
-    }
-
-    private void saveSuccessStatusToDatabase(Instance instance) {
-        Uri instanceDatabaseUri = Uri.withAppendedPath(InstanceProviderAPI.InstanceColumns.CONTENT_URI,
-                instance.getDatabaseId().toString());
-
-        ContentValues contentValues = new ContentValues();
-        contentValues.put(InstanceProviderAPI.InstanceColumns.STATUS, InstanceProviderAPI.STATUS_SUBMITTED);
-        Collect.getInstance().getContentResolver().update(instanceDatabaseUri, contentValues, null, null);
-    }
-
-    private void saveFailedStatusToDatabase(Instance instance) {
-        Uri instanceDatabaseUri = Uri.withAppendedPath(InstanceProviderAPI.InstanceColumns.CONTENT_URI,
-                instance.getDatabaseId().toString());
-
-        ContentValues contentValues = new ContentValues();
-        contentValues.put(InstanceProviderAPI.InstanceColumns.STATUS, InstanceProviderAPI.STATUS_SUBMISSION_FAILED);
-        Collect.getInstance().getContentResolver().update(instanceDatabaseUri, contentValues, null, null);
-    }
-
-    /**
-     * Returns whether instances of the form specified should be auto-deleted after successful
-     * update.
-     *
-     * If the form explicitly sets the auto-delete property, then it overrides the preference.
-     */
-    public static boolean formShouldBeAutoDeleted(String jrFormId, boolean isAutoDeleteAppSettingEnabled) {
-        Cursor cursor = new FormsDao().getFormsCursorForFormId(jrFormId);
-        String autoDelete = null;
-        if (cursor != null && cursor.moveToFirst()) {
-            try {
-                int autoDeleteColumnIndex = cursor.getColumnIndex(AUTO_DELETE);
-                autoDelete = cursor.getString(autoDeleteColumnIndex);
-            } finally {
-                cursor.close();
-            }
-        }
-        return autoDelete == null ? isAutoDeleteAppSettingEnabled : Boolean.valueOf(autoDelete);
     }
 }
